requirements.txt
django==4.1.7
django-ckeditor==6.7.0
django-debug-toolbar==4.2.0
dadata==21.10.1
Pillow==10.1.0
django-crispy-forms==2.0
django-simple-captcha==0.5.20
django-jazzmin==2.6.0
django-ipware==5.0.1
django-nested-admin==4.0.2
requests==2.31.0
validators==0.28.3
environs==9.5.0
psycopg2-binary==2.9.9
asgiref==3.7.2
certifi==2023.7.22
debugpy==1.6.7.post1
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==42.0.0
defusedxml==0.7.1
daphne==4.1.0
channels==4.0.0
channels-redis==4.1.0
django-js-asset==2.1.0
idna==3.4
redis==4.6.0
oauthlib==3.2.2
python-slugify==8.0.1
python3-openid==3.2.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
urllib3==2.0.7
xhtml2pdf==0.2.16
agora-token-builder==1.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
gunicorn==21.2.0
django-rosetta==0.9.9
django-ace==1.32.4
django-multiselectfield==0.1.12
webdriver-manager==4.0.2
selenium==4.24.0
django-rosetta==0.9.9
openpyxl==3.1.5
django_grapesjs==0.0.7
numpy==2.1.3
pandas==2.2.3
python-dateutil==2.9.0.post0
celery==5.4.0
beautifulsoup4==4.12.3
django-celery-beat==2.7.0
aiohttp==3.9.5
aiortc==1.8.0
html2image==2.0.5
html2image==2.0.5
django-tinymce==4.1.0